decimal for stock quanity handled

+++ Stretchtec/database/migrations/2025_09_30_105719_create_stocks_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('reference_no')->unique();
             $table->string('shade');
-            $table->integer('qty_available');
+            $table->decimal('qty_available');
             $table->string('notes');
             $table->string('uom', 5);
             $table->timestamps();
