+++ Stretchtec/database/migrations/2025_09_30_155900_create_stores_table.php
@@ -16,8 +16,8 @@ return new class extends Migration {
             $table->string('prod_order_no')->nullable();
             $table->string('reference_no')->nullable();
             $table->string('shade')->nullable();
-            $table->integer('qty_available');
-            $table->integer('qty_allocated')->nullable();
+            $table->decimal('qty_available');
+            $table->decimal('qty_allocated')->nullable();
             $table->longText('reason_for_reject')->nullable();
             $table->string('assigned_by');
             $table->boolean('is_qty_assigned')->default(false);
